+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/archive/networks/CriticNetwork.py
