+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/a2c/utils.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 import os
 import numpy as np
 import tensorflow as tf
