+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/console_util.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 from __future__ import print_function
 from contextlib import contextmanager
 import numpy as np
