+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/distributions.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 import tensorflow as tf
 import numpy as np
 import ch.hslu.wipro.ddpg.new.common.tf_util as U
