+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/mpi_adam_optimizer.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 import numpy as np
 import tensorflow as tf
 from mpi4py import MPI
