+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/mpi_moments.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 from mpi4py import MPI
 import numpy as np
 from ch.hslu.wipro.ddpg.new.common import zipsame
