+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/mpi_running_mean_std.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 try:
     from mpi4py import MPI
 except ImportError:
