+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/mpi_util.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 from collections import defaultdict
 from mpi4py import MPI
 import os, numpy as np
