+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/running_mean_std.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 import tensorflow as tf
 import numpy as np
 from ch.hslu.wipro.ddpg.new.common.tf_util import get_session
