+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/tf_util.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 import joblib
 import numpy as np
 import tensorflow as tf  # pylint: ignore-module
