+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/memory.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
+################################################################
+"""
+
 import numpy as np
