+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/reward/position_on_ground_pitch_reward.py
@@ -1,9 +1,6 @@
-import numpy as np
-
 from ch.hslu.wipro.ddpg.reward import RewardMultipliers
 from ch.hslu.wipro.ddpg.reward.reward_interface import RewardInterface
 from ch.hslu.wipro.ddpg.reward.reward_state import RewardState
-from ch.hslu.wipro.fg.calc.calc_distance import DistCalc
 
 
 class PositionOnGroundPitchReward(RewardInterface):
@@ -15,10 +12,12 @@ class PositionOnGroundPitchReward(RewardInterface):
 
     def calculate_reward(self, props) -> (float, bool):
         # add vector to observation space
-        reward_to_return = 0
-        pitch_deg = props['pitch-deg']
+        # reward_to_return = 0
+        # pitch_deg = props['pitch-deg']
+
         elevator = 1 + props['elevator']
 
+        # Evaluate elevator to change the pitch, not the pitch itself
         reward_to_return = RewardMultipliers.ON_GROUND_ELEVATOR_REWARD * elevator
 
         return reward_to_return, False
