+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/reward/unused/on_ground_rewards.py
