+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/reward/unused/position_in_air_pitch_reward.py
