+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/reward/unused/position_over_ground_bearing_reward.py
