+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/spaces/Box.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI GYM https://github.com/openai/gym
+################################################################
+"""
+
 import numpy as np
 
 from ch.hslu.wipro.ddpg.spaces.Space import Space
