+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/spaces/dict_space.py
@@ -1,3 +1,9 @@
+"""
+################################################################
+FILE FROM OPEN AI GYM https://github.com/openai/gym
+################################################################
+"""
+
 from collections import OrderedDict
 
 from ch.hslu.wipro.ddpg.spaces.Space import Space
